MonotonicTime added.

+++ interface.go
@@ -28,13 +28,12 @@ package wiringpi
 
 /*
 #include <time.h>
-#include <stdint.h>
 
-uint64_t as_nanoseconds(struct timespec* ts) {
-    return ts->tv_sec * (uint64_t)1000000000L + ts->tv_nsec;
+unsigned long long as_nanoseconds(struct timespec* ts) {
+    return ts->tv_sec * (unsigned long long)1000000000L + ts->tv_nsec;
 }
 
-uint64_t monotonic_time() {
+unsigned long long monotonic_time() {
     struct timespec last_t;
     clock_gettime(CLOCK_MONOTONIC, &last_t);
     return as_nanoseconds(&last_t);
